Sort layout into sections and hand grid parsing to the cascade
Grid track lists, area maps and placements were the one place layout
read raw CSS text. They now parse in styles.ts, where every other value
does, and layout imports the four it asks for.

What remains reads in order: the predicates that classify a box, style
becoming a flex node, the box tree, invalidation, line breaking,
geometry, and the engine that drives them all.
